\encoding{latin1}

\name{multi.ts}

\docType{data}

\alias{multi.ts}

\title{Generated Multivariate Time Series}

\description{
This data matrix contains a 4-variate time series of length 500. It consists of two Blocks and two Doppler signals each overlaid by highly correlated bivariate Gaussian noise.
}

\usage{multi.ts}

\format{A (500x4)-matrix containing a 4-variate time series of length 500.}

\source{Data generated by means of the packages \code{wmtsa} and \code{MASS}.}

\keyword{datasets}
