\encoding{latin1}

\name{sizecorrection}

\docType{data}

\alias{sizecorrection}

\title{Bias correction factors for the robust scale estimators MAD, Sn, Qn, and LSH}

\description{
This matrix contains correction factors for the MAD, Sn, Qn, and LSH scale estimators to achieve unbiasedness under Gaussian noise.
}

\usage{sizecorrection}

\format{
A (31x4)-matrix containing correction factors for the MAD, Sn, Qn, and LSH scale estimators.
}

\source{The correction factors have been obtained by simulations.}

\keyword{datasets}
