\name{robfilter-package}
\alias{robfilter-package}
\alias{robfilter}
\docType{package}
\title{
\packageTitle{robfilter}
}
\description{
\packageDescription{robfilter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robfilter}
\packageIndices{robfilter}
\preformatted{
adore.filter            A Robust Adaptive Online Repeated Median Filter
                        for Univariate Time Series
const.Q                 Correction factors to achieve unbiasedness of 
                        the regression-free Q scale estimator                       
const                   Correction factors to achieve unbiasedness of
                        the Qn scale estimator
critvals                Critical Values for the RM Goodness of Fit Test
dfs                     Degrees of freedom for the SCARM test statistic
dr.filter               Deepest Regression (DR) filter
dw.filter               Robust Double Window Filtering Methods for
                        Univariate Time Series
hybrid.filter           Robust Hybrid Filtering Methods for Univariate
                        Time Series
lms.filter              Least Median of Squares (LMS) filter
lqd.filter              Least Quartile Difference (LQD) filter
lts.filter              Least Trimmed Squares (LTS) filter
madore.filter           A Robust Adaptive Online Filter for
                        Multivariate Time Series
med.filter              Median (MED) filter
multi.ts                Generated Multivariate Time Series
rm.filter               Repeated Median (RM) filter
robreg.filter           Robust Regression Filters for Univariate Time
                        Series
robust.filter           Robust Filtering Methods for Univariate Time
                        Series
scarm.filter            SCARM (Slope Comparing Adaptive Repeated Median)                        
var.n                   Variance of the Repeated Median slope estimator
wrm.filter              Weighted Repeated Median Filters for Univariate
                        Time Series
wrm.smooth              Weighted Repeated Median Smoothing
}
}
\author{
\packageAuthor{robfilter}

Maintainer: \packageMaintainer{robfilter}
}
